do not detect tests in build.rs
